Remove controllers spec, make opa deployment automatic
