+++ .circleci/config.yml
@@ -73,9 +73,9 @@ jobs:
         command: |
           docker tag \
             ${CIRCLE_PROJECT_REPONAME}:latest \
-            replicated/gatekeeper:0.0.2
+            replicated/gatekeeper:0.0.3
           docker login --username $DOCKERHUB_USER --password $DOCKERHUB_PASSWORD
-          docker push replicated/gatekeeper:0.0.2
+          docker push replicated/gatekeeper:0.0.3
 
 workflows:
   version: 2
