--- config/samples/controllers_v1alpha1_openpolicyagent.yaml
@@ -1,11 +1,0 @@
-apiVersion: controllers.replicated.com/v1alpha1
-kind: OpenPolicyAgent
-metadata:
-  labels:
-    controller-tools.k8s.io: "1.0"
-  name: openpolicyagent-sample
-spec:
-  name: gatekeeper
-  enabledFailureModes:
-    ignore: true
-    fail: false
