+++ config/samples/policies_v1alpha1_admissionpolicy-fail.yaml
@@ -5,8 +5,8 @@ metadata:
     controller-tools.k8s.io: "1.0"
   name: no-latest-tags
 spec:
-  name: latest
-  failurePolicy: Ignore
+  name: latest-fail
+  failurePolicy: Fail
   policy: |
     package kubernetes.admission
