+++ config/samples/policies_v1alpha1_admissionpolicy-ignore.yaml
@@ -1,0 +1,18 @@
+apiVersion: policies.replicated.com/v1alpha1
+kind: AdmissionPolicy
+metadata:
+  labels:
+    controller-tools.k8s.io: "1.0"
+  name: no-latest-tags
+spec:
+  name: latest-ignore
+  failurePolicy: Ignore
+  policy: |
+    package kubernetes.admission
+
+    deny[msg] {
+        input.request.kind.kind = "Pod"
+        input.request.operation = "CREATE"
+        endswith(input.request.object.spec.containers[_].image, ":latest")
+        msg = "pod contains image using latest tag"
+    }
