--- docs/gatekeeper-k8s/crd-opa.yaml
@@ -1,58 +1,0 @@
-#
-# Install the Open Policy Agent Operator
-#
-apiVersion: apiextensions.k8s.io/v1beta1
-kind: CustomResourceDefinition
-metadata:
-  labels:
-    controller-tools.k8s.io: "1.0"
-  name: openpolicyagents.controllers.replicated.com
-spec:
-  group: controllers.replicated.com
-  names:
-    kind: OpenPolicyAgent
-    plural: openpolicyagents
-  scope: Namespaced
-  subresources:
-    status: {}
-  validation:
-    openAPIV3Schema:
-      properties:
-        apiVersion:
-          type: string
-        kind:
-          type: string
-        metadata:
-          type: object
-        spec:
-          properties:
-            enabledFailureModes:
-              properties:
-                fail:
-                  type: boolean
-                ignore:
-                  type: boolean
-              required:
-              - ignore
-              - fail
-              type: object
-            name:
-              type: string
-          required:
-          - name
-          - enabledFailureModes
-          type: object
-        status:
-          properties:
-            mainPolicyDeployed:
-              type: boolean
-          required:
-          - mainPolicyDeployed
-          type: object
-  version: v1alpha1
-status:
-  acceptedNames:
-    kind: ""
-    plural: ""
-  conditions: []
-  storedVersions: []
