--- docs/gatekeeper-k8s/opa-gatekeeper.yaml
@@ -1,15 +1,0 @@
-#
-# Create an installation of OpenPolicyAgent named "opa", with
-# failurePolicy set to ignore only
-#
-apiVersion: controllers.replicated.com/v1alpha1
-kind: OpenPolicyAgent
-metadata:
-  labels:
-    controller-tools.k8s.io: "1.0"
-  name: opa
-spec:
-  name: gatekeeper
-  enabledFailureModes:
-    ignore: true
-    fail: false
