+++ docs/gatekeeper-k8s/statefulset-manager.yaml
@@ -21,7 +21,7 @@ spec:
       containers:
       - command:
         - /root/manager
-        image: replicated/gatekeeper:0.0.2
+        image: replicated/gatekeeper:0.0.3
         imagePullPolicy: Always
         name: manager
         env:
